Extract analyze project agent definition
Moved the `analyze_project_agent` definition from the
`Project.analyze_page_content` method to a dedicated module,
`core/agents/analyze_project_agent.py`.

This refactoring improves code organization by separating the agent's
definition from the model's method logic. The agent is now a
standalone object, enhancing potential reusability and making the
model code cleaner.
